add changes

+++ additionalEdit.php
@@ -5,8 +5,10 @@ $description = $_POST['description'];
 $status = isset($_POST['status']) ? 1 : 0;
 $id = $_POST['id'];
 
-$namePhoto = $_FILES['photo']['name'];
-$tmp_name = $_FILES['photo']['tmp_name'];
+if ($_FILES['photo']['size'] !== 0) {
+    $namePhoto = $_FILES['photo']['name'];
+    $tmp_name = $_FILES['photo']['tmp_name'];
+}
 
 $sqlPhoto = "SELECT photo FROM products WHERE id = ?";
 $stmt = $pdo->prepare($sqlPhoto);
@@ -14,9 +16,18 @@ $stmt->bindValue(1, $id);
 $stmt->execute();
 $path = $stmt->fetch(PDO::FETCH_ASSOC);
 
-$pathToPhoto = 'uploads/' . date('YmdHis') . $namePhoto;
+//echo '<pre>'; var_dump($path);
+//echo '<br>';
+//echo '<br>';
+//echo '<br>';
+//echo '<pre>'; var_dump($path['photo']);exit();
 
-move_uploaded_file($tmp_name, $pathToPhoto);
+if ($_FILES['photo']['size'] !== 0) {
+    $pathToPhoto = 'uploads/' . date('YmdHis') . $namePhoto;
+    move_uploaded_file($tmp_name, $pathToPhoto);
+} else {
+    $pathToPhoto = $path['photo'];
+}
 
 $sql = "UPDATE products SET name = ?, description = ?, status = ?, photo = ? WHERE id = ?";
 $statement = $pdo->prepare($sql);
@@ -26,5 +37,9 @@ $statement->bindValue(3, $status);
 $statement->bindValue(4, $pathToPhoto);
 $statement->bindValue(5, $id);
 $statement->execute();
+
 header('Location: index.php');
-unlink($path['photo']);
+
+if ($_FILES['photo']['size'] !== 0) {
+    unlink($path['photo']);
+}
